chg: subjects from tree instead of clinicaldata
The old way of using clinicaldata as a source for a subject list
would mean that only subjects with at least 1 data point would be
included.

Getting the subjects from the tree instead means that reports about
recruitment and subject creation will be more accurate. Also makes
it more reliable to use as a source for ETL tasks that use the
webservices.