wip: working on extracting multi-value data
previously, the multi-valued data was already exploded to rows
in the clinicaldata query. Since that's not done anymore it needs to
happen at the item_group query level. This is work in progress to get
the correct values out for a mutli-value column, along with their
correct value label for the (value, label) column pairs.